requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
yagmail==0.15.293
flask-cors==4.0.0
keyring==24.3.0
secure-smtplib==0.1.1
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0 